requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
google-adk[litellm]>=1.0.0
litellm>=1.50.0
deprecated>=1.2.0
